add catch-all not-found routes
