+++ src/app/[lang]/[...404]/page.tsx
@@ -1,0 +1,5 @@
+import { notFound } from "next/navigation";
+
+export default function page() {
+  notFound();
+}
