+++ src/app/[lang]/articles/error.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+export default function Error({
+  error,
+  reset,
+}: {
+  error: Error;
+  reset: () => void;
+}) {
+  return <div>{error.message}</div>;
+}
