+++ src/app/[lang]/articles/not-found.tsx
@@ -1,0 +1,5 @@
+import PageError404 from "@/components/404/404";
+
+export default function NotFound() {
+  return <PageError404 />;
+}
