+++ src/app/[lang]/articles/page.tsx
@@ -3,7 +3,8 @@ import { Locale } from "@/i18n-config";
 import ArticlesList from "@/components/articles/ArticlesList";
 import CardArticle from "@/components/articles/CardArticle";
 
-import { getAllArticles } from "@/services/services";
+import { getAllArticles } from "@/services/api-client";
+import { notFound } from "next/navigation";
 
 export default async function Articles({
   params: { lang },
