+++ src/app/[lang]/global-error.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+export default function errorPage({
+  error,
+  reset,
+}: {
+  error: Error;
+  reset: () => void;
+}) {
+  return <div>{error.message}</div>;
+}
