--- src/services/apiClient.ts
@@ -1,1 +1,0 @@
-export class ApiClient {}
